+++ CHANGELOG.md
@@ -30,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added possibility of creating basic systems in lua scripts (#1518, **@mkuritsu**)
 - Sane default placement for tools in tesseratos (#1387, **@jdbaracho**).
 - Layout structure with loading and ImGui application (#1387, **@jdbaracho**).
+- Allow generating metas with random UUIDs through Quadrados (#1441, **@Fkatar**)
 
 ### Changed